+++ src/Ext/Techno/Body.h
@@ -56,6 +56,7 @@ public:
 		CellClass* FiringObstacleCell; // Set on firing if there is an obstacle cell between target and techno, used for updating WaveClass target etc.
 		bool IsDetachingForCloak; // Used for checking animation detaching, set to true before calling Detach_All() on techno when this anim is attached to and to false after when cloaking only.
 		int BeControlledThreatFrame;
+		int LastHurtFrame;
 		DWORD LastTargetID;
 		int AccumulatedGattlingValue;
 		bool ShouldUpdateGattlingValue;
@@ -103,6 +104,7 @@ public:
 			, FiringObstacleCell {}
 			, IsDetachingForCloak { false }
 			, BeControlledThreatFrame { 0 }
+			, LastHurtFrame { 0 }
 			, LastTargetID { 0xFFFFFFFF }
 			, AccumulatedGattlingValue { 0 }
 			, ShouldUpdateGattlingValue { false }