+++ src/Ext/TechnoType/Body.h
@@ -227,6 +227,11 @@ public:
 
 		Nullable<bool> RecountBurst;
 
+		Valueable<double> Skilled_ReverseSpeed;
+		Valueable<double> Skilled_FaceTargetRange;
+		Valueable<bool> Skilled_ConfrontEnemies;
+		Valueable<int> Skilled_RetreatDuration;
+
 		ValueableVector<TechnoTypeClass*> BuildLimitGroup_Types;
 		ValueableVector<int> BuildLimitGroup_Nums;
 		Valueable<int> BuildLimitGroup_Factor;
@@ -508,6 +513,11 @@ public:
 			, AttachEffects {}
 
 			, RecountBurst {}
+
+			, Skilled_ReverseSpeed { 0.85 }
+			, Skilled_FaceTargetRange { 16.0 }
+			, Skilled_ConfrontEnemies { true }
+			, Skilled_RetreatDuration { 150 }
 
 			, BuildLimitGroup_Types {}
 			, BuildLimitGroup_Nums {}