+++ docs/Whats-New.md
@@ -366,6 +366,7 @@ Vanilla fixes:
 - Fixed the bug that `EnterBioReactorSound`, `LeaveBioReactorSound`, `EnterGrinderSound` on technotype does not used (by NetsuNegi)
 - Fixed the bug that harvester dont stop unloading and cannot unload cargos anymore when lifting by `IsLocomotor=yes` warhead (by NetsuNegi)
 - Fixed an issue that units on the slope tilted at an excessive angle (by CrimRecya & NetsuNegi)
+- New Skilled locomotor (by CrimRecya)
 
 Fixes / interactions with other extensions:
 - Allowed `AuxBuilding` and Ares' `SW.Aux/NegBuildings` to count building upgrades (by Ollerus)